fixed auth and added auth middleware

+++ src/modules/commentModule.js
@@ -1,5 +1,8 @@
 import commentRepo from '../repo/commentRepo.js';
 import commentSchema from '../schema/commentSchema.js';
+import mongo from '../service/mongo.js';
+import { ObjectId } from 'mongodb';
+const getCollection = mongo.getCollection;
 
 const create = async (ctx) => {
     const { value, error } = commentSchema.validate(ctx.request.body);
@@ -8,7 +11,7 @@ const create = async (ctx) => {
         return (ctx.body = error.details[0].message);
     }
 
-    value.userId = 1;
+    value.userId = ctx.state.authData.user._id;
 
     const insertedId = await commentRepo.createComment(value);
     return (ctx.body = {
@@ -18,10 +21,24 @@ const create = async (ctx) => {
 };
 
 const remove = async (ctx) => {
-    commentRepo.removeComment(ctx.params.id);
-    return (ctx.body = {
-        success: true,
+    const comment = await getCollection('test', 'comment').findOne({
+        userId: ctx.state.authData.user._id,
+        _id: new ObjectId(ctx.params.id),
     });
+
+    if (comment) {
+        commentRepo.removeComment(ctx.params.id);
+        return (ctx.body = {
+            success: true,
+            message: 'comment deleted successfully',
+        });
+    } else {
+        ctx.status = 403;
+        return (ctx.body = {
+            success: false,
+            message: 'you dont have prermission to delete this comment',
+        });
+    }
 };
 
 export default { create, remove };
